[SRP] Fixing Lens Flare Screen Space And Lens Flare Data Driven failures on yamato
Following the merge of Screen Space Lens Flare feature, a PS4 image update has been overlooked. This fixes this issue. 
This removes the need from backbuffer for the lens Flare Sanity tests which was causing Metal to fail as well. 
This tests also adds correct meta file with proper settings for mobile testing: 
- MipMaps: Disabled
- Read/Write: Disabled
- Compression: None