[Universal] Release non-active renderer RTs
In non-render graph path, if the project has multiple renderers and goes through them - we do not dispose until pretty much exiting the application. This is especially hurting to our URP tests, which use massive amount of renderers. This shouldn't affect user projects too much, as they usually contain one or two renderers that are actually used. 

This also fixes yamato OOM issue for one of the runners