[VFX Graph] Add usage analytics
[Jira](https://jira.unity3d.com/browse/VFXG-176)

The goal of the PR is to start monitoring VFX Graph usage on a few user interaction.
Here is the list of interactions that we want to monitor:
- How many unique users use VFX Graph (users who opened at least once the VFX Graph)
- Number of VFX assets in projects
- Number of nodes in VFX Graphs
- Usage of experimental features
- Usage of samples / helper additional samples
- Usage of pseudo templates (Create Node => Systems)
- Usage of custom shaders in Output contexts
- VFX Graph compute shader compilation errors.

[Doc](https://docs.google.com/document/d/1po6dBwtnI-B8kehmiaQgrU1QQVxyrnR4Yqq3jYKhw2o/edit?usp=sharing)

🗨 Note
Some metrics would be nice to have per user (experimental nodes, templates and errors) but the way Google Data Studio handle "array field" prevent us to do it. Unless we always send all possible values with a flag telling if it was used or not. I think this approach is not really maintainable. So instead we can simply compare values between themselves (which is most used template, which is the most used experimental feature or which compilation error happens the most often).