[SG] Fix UI memory leaks in ShaderGraph
Details of work covered in this bug: https://jira.unity3d.com/browse/UUM-18995

Currently when the ShaderGraph window is opened and closed, the RAM consumed by Unity will increase a permanent amount for that session of the editor due to the Shader Graph tool not properly releasing handles to resources, preventing the garbage collector from kicking in and deallocating this memory, and causing what is known as a memory leak, as a result.

It should be noted that this PR does not fix ALL the memory leak issues, there are still a number of view elements that do not properly dispose of their elements, but this PR addresses the worst offenders. There are some underlying issues in the Graphview API itself that the ShaderGraph tool is built on that are harder to fix due to requiring changes in that now deprecated package, but this fixes a number of issues that can be addressed on the ShaderGraph side of things, and sets up rigging in order for subsequent PRs to do more cleanup across the board in the future. 

Video of issue: 

https://media.github.cds.internal.unity3d.com/user/1946/files/92242dd6-9458-451e-b748-62702172c6b7

Note the memory allocation counter at the bottom right of the screen, you'll see that every time a ShaderGraph is opened up, the memory allocations increase by ~15-20 MB, and then when the window is closed, ~5 MB gets released, leaving ~10-15 MB of garbage that cannot be deallocated.

**Before PR:**

Hanging visual elements:

![VisualElementsHangingAround](https://media.github.cds.internal.unity3d.com/user/1946/files/5eceb2f9-4a0b-497d-828a-f83b549d0b88)

Hanging node references:

![SubgraphAsset](https://media.github.cds.internal.unity3d.com/user/1946/files/7e2cc00e-62ed-4208-a8a8-57a50f248877)

![BeforePR_NodeLeaks](https://media.github.cds.internal.unity3d.com/user/1946/files/be6e730e-fc1d-4ee3-828b-281cc0d48934)

The majority of these leaks come from the UI classes not properly clearing references and unbinding callbacks at the time of the window close. 

**After PR:**

https://media.github.cds.internal.unity3d.com/user/1946/files/628258e3-b4dd-49fc-afe0-6cc1f1235793

Note the memory allocation on the bottom-right counter, after the PR is applied, the memory allocation when a graph is closed falls by ~10 MB as opposed to ~5 MB before the PR. Of the ~15-20 MB that is allocated when a graph is opened, we are releasing slightly more than half now as opposed to a third of the memory that we were before. 

Subgraph nodes don't leak anymore, only instance is the default object that Unity creates in memory for all unity objects it looks like. 

![SubgraphAssetReferenceCleared](https://media.github.cds.internal.unity3d.com/user/1946/files/2aeffba9-ea86-4159-9337-39e3cb60b598)

There's still a lot of garbage leftover, but the garbage associated with each node that was created on the graph is no longer present.

![image](https://media.github.cds.internal.unity3d.com/user/1946/files/5443a130-dac6-43f6-b42e-5245ba6c0cc1)