[RPW] Fix Texture2D leak from TextureCurveParameter when changing scenes inside a Volume.
Fix for https://jira.unity3d.com/browse/UUM-16686

The bug happens because `TextureCurve` has an implementation where it will lazily create a `Texture2D` object when it is needed, in practice this happens when the `ColorGrading` object from the VolumeStack is accessed and used by the requested by ColorGradingLutPass. However, due to how `Interp` is implemented, the VolumeStack only contains a _reference_ to a `TextureCurveParameter`, which itself is owned by a `VolumeProfile` (a `ScriptableObject`). The end result is that during rendering, we have created a `Texture2D` object inside one of the `VolumeProfiles` referenced by the Volumes in the scene. When the scene is unloaded, the `VolumeProfiles` or `VolumeComponents` do not get destroyed because as ScriptableObjects, their lifetime is longer than the lifetime of GameObjects. Instead, Unity calls `OnDisable` on them after the last reference to them goes out of scope. Therefore we call `Release` at this point to release the runtime resources (`Texture2D` in this case) and avoid the leak.