[VFX][Fix 1361731] Baked Gradient and Curve overflow
When high values ( > 65504.0f) are present in a Gradient or a Curve evaluated on the GPU, sampling the bakedTexture returns INF. That is because the bakedTexture is stored with half-float channels.
This PR works around this issue by scaling down, when necessary, all the values in the bakedTexture so they don't overflow. An additionnal multiplier is sent to the GPU to inverse this scaling.
For gradients, an additionnal float is included in the gradient parameters (before, the parameters were a float2, now it's a float3).
For curves, as the parameters were already a float4, the first two parameters have been repacked as half floats to make room to the multiplier.

Jira Bug: https://jira.unity3d.com/browse/UUM-4305