SRP CullShadowCasters API
**Epic:** https://jira.unity3d.com/browse/PLAT-1168
**POI:** https://jira.unity3d.com/browse/POI-467

Currently, the BatchRendererGroup (BRG) visibility culling is handled using the OnPerformCulling callback. While the camera culling is invoked in `ScriptableRenderContext.Cull`, shadow caster culling is invoked in `ScriptableRenderContext.Submit`. This is very late in the frame and the shadow culling jobs end up being kicked too late which causes the main thread to wait idle.

This PR tries to improve this by add a new API `ScriptableRenderContext.CullShadowCasters` which invokes the OnPerformCulling callbacks for shadow casters culling. In a future PR, it would be great to also kick GameObjects shadow casters culling here. Especially `ComputeShadowSplitMasks` in `PrepareDrawShadowsCommandStep2`.