[VFX] Add shader debug symbols toggle
This PR adds a way to enable shader debug symbols in VFX Graph generated shaders.

This is useful during development to get more debug info in shader debugging tools like render doc and allow fine debugging of shaders.

This PR adds one general toggle in VFX preferences to affect all VFX locally
<img width="588" alt="sds2" src="https://media.github.cds.internal.unity3d.com/user/346/files/38e727c6-1bc4-4f70-ba1e-af63dba3314f">

It also adds a toggle in VFX Graph view to allow toggling it for a single VFX
<img width="205" alt="sds1" src="https://media.github.cds.internal.unity3d.com/user/346/files/e237044a-dda4-4b80-ae91-c98bc7575774">
Note that this toggle is in the view context, it is not serialized in the asset. Meaning that shader symbols are added only when the VFX is authored in view. When the VFX is closed, it is recompiled using the global state defined in preferences.

The change is only applied upon VFX recompilation. A pop up was added when toggling it globally in VFX preferences to ask users if they want to recompile all (as this may take some time). The pop up also shows when toggling "Force Compilation in Edit Mode". Before all VFX were recompiled without notice.
<img width="293" alt="sds3" src="https://media.github.cds.internal.unity3d.com/user/346/files/b996186d-7307-4e8e-9610-e42166b2eeda">

https://jira.unity3d.com/browse/UUM-19519