[2022.2][SpeedTree][URP][HDRP] Multiply the slider value for smoothness and metallic with the texture data of RP shadergraphs
https://jira.unity3d.com/browse/ENV-3294

![image](https://media.github.cds.internal.unity3d.com/user/5527/files/f5f24427-6221-4aec-bc37-d6d2ec6b0c3d)

The smoothness and metallic slider are only active when there's no smoothness and/or metallic textures used. The expected behavior is both of them working together if the sliders are visible. Other importers either don't expose the slider, or if they do, the slider value is multiplied by the texture values to calculate effective smoothness value the shader will eventually use for lighting. 

This PR changes the shadergraph shaders to multiply Smoothness and Metallic with the ExtraTexture when the texture is provided while the Importer sets the default value of Smoothness and Metallic to `1.0f` when there's an ExtraTexture is present during importing.

Also changed is the URP's SSS logic for the Smoothness input.

It used to read the slider value regardless of the ExtraTexture values, now a single smoothness value is used to calculate both SSS contribution and direct lighting.

**BEFORE**

URP
![image](https://media.github.cds.internal.unity3d.com/user/5527/files/2399353f-0bc0-4317-9532-0977ba1a0b2e)
![image](https://media.github.cds.internal.unity3d.com/user/5527/files/09cb9f55-41d5-4787-8481-5d997f8a66cf)


HDRP
![image](https://media.github.cds.internal.unity3d.com/user/5527/files/833c506d-1082-434c-9d06-b89d3611cd65)


**AFTER**

URP
![image](https://media.github.cds.internal.unity3d.com/user/5527/files/46f2b04c-6fa6-48f6-8800-53293d21cb97)

HDRP
![image](https://media.github.cds.internal.unity3d.com/user/5527/files/9b0afdbb-15e6-4db4-8d58-24c13f9ac26e)