[SRP] Move Lens Flare Data Driven Post Process after Bloom (2022.2)
This recently merged PR (#16831) introduces Screen Space Lens Flare. 
Since it's based on the bloom buffer, we needed to remove the Data Driven Lens Flare from it otherwise, it would create lens flare of lens flare and that would be a bit too much. So eventhough it made sense physically speaking, light bouncing in the lens and then getting diffracted in the opening of the camera (bloom), the order needed to change. 

Before: 
- Lens Flare Data Driven
- Bloom Pass
- Uber

After:
- Bloom Pass
- Lens Flare Screen Space
- Lens Flare Data Driven
- Uber

Since this alter slightly the look of some scenes, this PR changes also the order in 2022.2 (and same will be done for 2022.1) eventhough there's no Screen Space Lens Flare, to avoid "breaking" more project. 

The consequence of that is in 2023.1, in scenes where there's Bloom, Lens Flare Data Driven looks a bit "dryer" since they aren't on screen when bloom is applied. Here's a preview of the change in the HDRP template:

![2d28b8df5072a52e9d7ff2314dd18df1](https://media.github.cds.internal.unity3d.com/user/1764/files/df4b77b7-fafd-49d9-8d02-6a4202554c5e)