Gsg 1440
This allows graph shaders to set the disableBatching tag, which is necessary to have a shader opt out of dynamic batching.  For existing targets that exposed a LOD Crossfade option I wired that up to set the tag to LODFading which should be the expected behavior.  This work was done in response to https://jira.unity3d.com/browse/SGB-299 and https://jira.unity3d.com/browse/GSG-1440 and serves as foundational work to eventually fix https://jira.unity3d.com/browse/UUM-12488.