[HDRP] Fix forward only material validation
Render Pipeline tag is sometimes not setup yet during the Asset PostProcessor callback, which results in wrong subshader selection during material validation.
This is bad, because we can't do things such as get pass count and name during validation as it doesn't always select the HDRP subshader.

The fix changes the way we detect forward only materials to avoid the issue.