[VFX] Fix unsafe deletion of VFXParticleSystem
Fix unsafe deletion of VFXParticleSystem & VFXMeshSystem copying needed data in VFXRenderCommand. The problem was visible on a repro project on Switch:
![image](https://media.github.cds.internal.unity3d.com/user/42/files/d29bf6ca-6337-49a2-bd1a-9adac870d254)
As exposed ⬆️ the problem was about the suppression of the VFXParticleSystem between culling and actual rendering.
I wasn't able to reproduce somewhere else but it wasn't a platform specific issue.