[RPW] Add RPCore_PerformanceTests project
Add a new RPCore_Performance test project that will contain performance tests for RP Core functionality, such as VolumeManager. The intent is not to add this to any nightly CI right now, but we should discuss if we want to do it in the future. The immediate need for this project is to be able to measure performance impact RP Core package changes either locally or with Bokken.

Some features of the project:

- The project is defined similar to HDRP_PerformanceTests project, except it ~contains both URP and HDRP package~ only contains RP Core package.
- ~The active pipeline is set to HDRP.~
  - ~Ideally in the future we would have a way to run player build performance tests in CI using either HDRP or URP, but AFAIK this is not possible today.~
- There are two test assemblies, `Unity.GraphicTests.Performance.RPCore.Editor` and `Unity.GraphicTests.Performance.RPCore.Runtime` that will contain the performance tests. These assemblies can be used to expose internals of SRP packages for testing as needed (using `InternalsVisibleTo`). 
- The `com.unity.testing.graphics-performance` package is *not* used right now. Since concrete render pipelines are not included, probably there will be no need for actual graphics perf tests in this project.
- Manifest contains `"disableProjectUpdate": true` and `"enableLockFile": false`.
- .buginfo is configured to RP Workflows.