Use serialized shader data to find the tags instead of editor's runti…
Fixes: https://jira.unity3d.com/browse/UUM-14800
BiRP shadergraph stripping was hitting exceptions during Viking Village perf test project builds on Vulkan and thus failing the builds. It turned out that the stripping code was trying to get tags from the editor runtime instance of the shader using the serialized data indices. Fixed the code to get the tags from the serialized shader data instead. This is a backport PR