[HDRP] Fix migration step for Amethyst feature that have a new Frame Setting
Following Amethyst new HDRP features, there's a bunch of frame settings that were added. 
By default those Frame Settings used the default int value, which is 0, thus they were false. 
There were only true in the case where a user create a new HDGlobalSettings, which is a very low % of users. 
This led to a lot of users having trouble using those features. 
To avoid this, this migration step will set those to true by default on projects upgrade.
This fix should take care of the templates update which is the start of many HDRP users I believe.

This includes:
- Water
- WaterDeformation
- WaterExclusion
- ComputeThickness
- LensFlareScreenSpace
- HighQualityLineRendering
- VFXRaytracing