[2021.3] Fix float16/float32 errors in URP and BatchRendererGroup
On iOS, and on Mac in iOS mode (Build platform set to iOS), the Unity shader compiler will compile real4 and half4 in cbuffers to actual 16-bit floats, unlike on other platforms where they remain 32-bit in the cbuffer. This causes a problem:
- unity_ProbesOcclusion being half4 makes all URP shaders SRP Batcher incompatible, which is a hard error for BatchRendererGroup and a serious performance regression for GameObjects.

This PR reverts the use of half4/real4 for this constant, so 32-bit floats are consistently used on all platforms for these values. This can potentially be a small GPU performance regression, but it's required for SRP batcher compatibility and for BatchRendererGroup to work at all on iOS.

The trunk PR also fixes a similar issue in BatchRendererGroup, but that fix does not apply to 2021.3 because the shader code is different.

Fixes: https://jira.unity3d.com/browse/UUM-21379

Trunk PR: https://ono.unity3d.com/unity/unity/pull-request/147058/_/graphics/brg-fix-ios-float16