Fix subdivision levels not been displayed after baking with higher subdiv
Fix for  https://jira.unity3d.com/browse/UUM-20112 

was a bit harder to track down than I wished. 


The issue was that the debug max subdiv slider was clamped to a lower max subdiv in previous iterations, but as the max subdiv increases the debug is left to an older value.