Skip processing scenes without a profile when in Single scene mode
If we have two scenes loaded but in single scene mode, it is a valid situation that one of the two loaded scene has never been put to a set and so it doesn't have one. 

In those situations we just go over them; this will be ineherently fixed by future changes from @julien anyway (where we'll loop through stuff in sets). 

Fix for https://jira.unity3d.com/browse/UUM-20870 

Now the active scene bakes as expected