Allow bigger threshold for validity of the reflection probe normalization data
Fix for https://jira.unity3d.com/browse/UUM-18953 

Before this, if even a single backface was seen, then the normalization data would have been considered invalid. With increasing amount of samples, the likelihood of hitting at least a single backface increases and the weird behaviour happens. 

With this PR we are a bit more allowing. 

**NOTE: @maxime-grange  will verify on his end too, please wait for him before merging**