Automatically release leaked BatchRendererGroups and don't repeat the error messages
When Unity does a domain reload (e.g. when the user changes script code), it's possible for some objects to leak if scripts are not written correctly to handle the situation. If this happens to a BatchRendererGroup object, Unity will keep repeatedly issuing a loud error message that the user can only remove by restarting the editor, which is very disruptive.

This PR turns the error message to a warning, and will automatically release the leaked object when detected, so the warning is issued only once and does not severely impact editor usability.