Avoid importing all assets on project open
Fix performance regression introduced by https://github.cds.internal.unity3d.com/unity/unity/pull/9145.
It caused a lot of unecessary import of assets to check if they contain a material.

This was done to handle the case of assets containing material subassets but it's very niche so not worth it. The main use case would be to handle shadergraphs, but diffusion profile serialization has apparently always been broken inside shaders so it actually can't be used in that case.

I reverted the postprocessor code to how it was before, so auto registration only happens on material assets now.