Added option to disable clouds in the effects menu
The effects menu inside the scene view allows to enable and disable specific rendering features in the scene. In this PR an additional option for the clouds is being added. This is only exposed if the SRP supports clouds. An example of the usage can be found [here](https://drive.google.com/file/d/1kaEf_xUssfCA5sGWbgCXuJN4VyMmswc1/view?usp=share_link).