Fix for 2021.3.X: Terrain is not affected by Environmental Lighting when Draw Instanced is turned on
When the `Draw Instanced` terrain setting is enabled, Terrains were not taking ambient environment color lighting into account in URP projects. This is due to an `EVALUATE_VERTEX_SH` define that was introduced for URP to only sample spherical harmonics (where the environment lighting was stored) per vertex when needed -- terrain was not appropriately passing it's vertexSH value to the SH sampling function in the instanced normal path.

In trunk this bug was not manifesting itself due to another URP bug hiding it: the `EVALUATE_VERTEX_SH` define is never set in trunk, so no matter what the terrain (and other shaders) are always sampling per pixel, and so this buggy code path is not taken. Another bug fix should be submitted to fix this for trunk by the URP team, and I will ask them to include this fix when they do.  

BEFORE:
![before](https://media.github.cds.internal.unity3d.com/user/4225/files/ff359f02-95f6-408c-ba1b-6a5072bedd1b)

AFTER:
![after](https://media.github.cds.internal.unity3d.com/user/4225/files/f5839286-72b1-4701-81e4-b334dec900aa)