[21.3] Backport #18609 - Call Base GTF class from the HDRP Test Runner
The HDRP Runtime Test project is building all graphics API image bundles on windows, even if the vulkan argument is passed for example. This is because the hdrp testing package is defining an override of the graphics test framework SetupProject.ApplySettings method, and the "base" GTF method is never called. We therefore never iterate over the graphics APIs and cannot specifically request only a bundle to be built (e.g. the vulkan one).

This resulted in the vulkan tests relying on the win dx11 reference images (obviously incorrect behavior).

This PR is fixing it by calling the GTF method from inside the HDRP testing package. To do so we need to add the GTF reference to the HDRP testing asmdef.

See this slack thread for more details: https://unity.slack.com/archives/CU41BKSJF/p1667921748281349

Also took the opportunity of this PR to fix the NDA verification check trigger on 21.3 that was wrongly copy/pasted from the 22.2 backport.