Y flip in Copy Depth
This is a fix to yet another depth y-flip bug.
A big issue in CopyDepth is that `renderingData.cameraData.IsCameraProjectionMatrixFlipped()` checks `renderer.cameraColorTargetHandle` in copy depth pass, then puts the result as isSourceYFlipped.
What the code is doing is checking **color target of the renderer** to set a bool on a **source of blit of depth**.

RCA: The way our code is structured, it far too complicated to properly know when to perform a y-flip.
The entire logic is centered around the action of flipping with all the cases coming in to tell when it must be done.
In reality the problem is a problem of transition from y-up space to y-down space and vice versa.

The issue now becomes more simple:
```
performYFlip = IsFlipped(source) != IsFlipped(destination)
```

This PR is my attempt at simplifying the logic in this way. I initially put the `IsFlipped` function as static in `CopyDepth` to fix a specific bug, but this might be something to make available in `RenderingUtils`. Maybe we already have a function like this.