[URP] Reverted RenderSingleCamera deprecation from error to warning
As part of the 23.1 deprecation update cycle, we bumped the RenderSingleCamera deprecation state from warning to error.
After consultation, we realized that because of missing upgrade documentations and the fact that all the URP samples haven't been upgraded yet, the deprecation state was bumped by mistake and needs to be reverted to warning.

Not reverting it, user projects and URP samples will be in a broken state