Do not show URP Tonemapping Volume warning when mode is None
Preparation for https://jira.unity3d.com/browse/XPIPELINE-228

Fixes a minor issue where URP TonemappingEditor shows a warning about "tonemapping being used", even if the selected mode is "None" (i.e., tonemapping is not used). This is not a major issue right now, but once Default Volume Profile for URP is added in URP Global Settings, this warning would always be visible when HDR is unchecked in RP Asset.

Before this PR:
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/d69f1bb0-88f2-4098-9b1e-7ab553f66466)

After this PR: 
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/c3f19195-f450-43b9-8105-dbdb48592976)