[VFX] Culling issues (new PR)
This PR fixes an issue with command culling in VFX Graph. Prior to this work, the shadows culling results were not taken into account in the VFX commands culling, causing issues like shadows flickering.
The main issue was that the shadow culling callbacks were sent to the VFX Manager too late, after it already added commands to the current command buffer.

With this PR, the shadow culling callbacks are now doing nothing, and the gathering of the VFX renderers visible by lights is done only once, during VFX.ProcessCameraCommand.

No changes were necessary in HDRP. In URP, the only change was to execute VFX.ProcessCameraCommand earlier, before the shadows are rendered. This part of the fix had already been discussed and proposed [here](https://github.cds.internal.unity3d.com/unity/unity/pull/13861), but we did not merge it because I prefered to have it as part of this current PR.

In addition, several fixes were made : 
- The lifetime of a VFXCullResults is now tied to its corresponding CullResults : it is destroyed inside ~CullResults(). This required making an interface for the VFXManager.
- Fixing this lifetime issue also allows us to reenable the VFXCullJob to run multithreaded, and to avoid copying data into the JobData.
- The RenderMultiple path was causing out of bounds issues due to the batchBegin being taken into account twice.
- In the case of an update with multiple steps, m_InstanceIndices was empty, causing errors in the VFX culling job.
- activeInstanceCount was being used instead currentInstanceCount in the compute commands.