Add support for rendererPriority to BRG
This PR adds support for rendererPriority to BatchRendererGroup, so that scenes that rely on this feature (such as hair shading tests in HDRP) can render correctly:
- 1220_Sorting
- 1401_HairGraph_Area_Light
- 8102_Transparent

This feature is required for parity with game object rendering, discovered as part of the testing done for Tier 0 of GPU-Driven Rendering (#15328).  Since BatchRendererGroup is also used by Entities Graphics, it can also be used by Entities Graphics to achieve parity with game object rendering.

After discussion with the BRG team, this has been implemented by adding support at the BatchDrawRange level, to minimise the impact on memory usage BRG structures. This is backwards compatible with existing code using BRG, which will continue to have the default priority value of 0.