URP: Remove support for Deprecated non-RTHandle use in ScriptableRenderer.SetRenderPassAttachments
RTHandles support for URP was added in 2021.3. In order to make transition possible, a compatibility layer was added to support the now-obsolete `RenderTargetHandle` with the use of `GetTemporaryRT` and the use of `RenderTargetIdentifier` was replaced with `RTHandles` in several `ScriptableRenderer` functions.

An upgrade guide was written for devs and users and the relevant functions were marked as obsolete.

Now that we're in 2023.2, this PR removes this compatibility layer and reduces the complexity associated with supporting both paths.