[VFX] Performance improvements on asset importing and fixed compilation issue on OSXe
https://jira.unity3d.com/browse/UUM-21840
https://jira.unity3d.com/browse/UUM-21618

Original PR: https://github.cds.internal.unity3d.com/unity/unity/pull/20382

This PR fixes two issues:
**UUM-21840**
When importing any asset, all asset post processors are called.
One of the VFXGraph post processor was systematically doing a file access (which is a slow operation) even when it could be avoided.

- Now this file access is performed only on MacOS (because it's used to fix an issue on MacOS only) and only VFX Graph assets.
- In addition some string comparison operations could be done a bit faster.

**UUM-21618**
Fixed a compilation issue on MacOS due to an unexpected reference to a .NET namespace that is not supported on MacOS.