[2023.2] Remove #ifdefs from UnityPerMaterial property definitions to avoid SRP Batcher incompatibility
A recent Shader Graph change modified how UnityPerMaterial cbuffers were generated for Shader Graphs. Unfortunately, it was not noticed in the review that this caused an #ifdef to end up in the UnityPerMaterial cbuffer definition. This causes the shader to be incompatible with the SRP Batcher, which is a big performance regression for GameObjects, and completely breaks BatchRendererGroup.

This PR causes the cbuffer to unconditionally generate the constants in UnityPerMaterial, so this problem doesn't occur.