[RPW] Refactor VolumeProfileCreator to Core
Preparation work for https://jira.unity3d.com/browse/XPIPELINE-228.

In preparation for adding a Default Volume Profile for URP, I'm moving VolumeProfileCreator functionality to core. The PR is intended to have no functional changes, just a refactor. At the same time, refactored `CreateParentFolders` to a separate function and added tests for it.