Support reading sampled unity_ProbesOcclusion from DOTS instanced SH Coefficients
Prior to this PR, unity_ProbesOcclusion is treated as a separate DOTS instanced property, which is not set in practice by any Unity-supplied code. This does not reflect its actual usage, which is to provide cheap baked shadowing data for Mixed lights in some GI modes (e.g. Subtractive).

This PR combines the occlusion probe sample into the already existing DOTS instanced SH coefficient struct, so it can be conveniently sampled in the same place where the normal SH coefficients are sampled. This is also convenient, because the SH struct contained one float4 of padding, and the occlusion probe sample can just occupy that same space without increasing the size of the struct.

This PR makes it possible for BatchRendererGroup based renderers (e.g. Entities Graphics and the GPU driven renderer) to implement blended occlusion probe sampling efficiently, which is required for correctly supporting Mixed Lights in Subtractive mode (and possibly other combinations).