[HDRP] Fix box-shape spotlight culling planes when used with BRG
This PR updates the culling planes for box-shaped spotlights to be consistent with their volume of influence.  Currently box-shaped spotlights have culling planes that are still in a pyramid-shape, which can lead to missing shadow casters when used with BatchRendererGroup.

The issue does not occur with game objects due to some SRP code that overrides the culling planes for spot lights, but it is not expected that users that implement the BRG culling callback should have to implement similar overrides.

This issue was originally found in SpaceshipDemo when rendering using BRG.  The correct shadows with game objects are as follows:
![image](https://media.github.cds.internal.unity3d.com/user/4787/files/7ee2dec2-5fe4-41f8-9c3b-2db58d84bca6)

Currently, when rendered with BRG, shadows are missing:
![image](https://media.github.cds.internal.unity3d.com/user/4787/files/10150917-2c14-41bd-9f26-8b16dc7fbd39)

Due planes be overridden for game objects, these changes only affect rendering with BRG.