[VFX] Fix VFX instance leak due to asset editor not properly destroyed
[Fix of UUM-25712](https://jira.unity3d.com/browse/UUM-25712).
When dragging and dropping a Visual Effect asset to the scene, game objects are silently created and never destroyed.
This is only noticeable when looking at the allocated memory, but internally it also affects performance, since it adds new empty VFX that need to be checked for update every frame.
There are 2 parts for the fix in this PR:
- Under some circumstances (asset editor not having a custom OnSceneDrag) we are creating 2 asset editors and only destroying 1. The fix for this is adding the missing DestroyImmediate for the second editor.
- On VFX side, even after the first fix, we were creating 1 VFX for each Visual Effect asset editor, to be used in the preview window. Even if those are destroyed after drag and drop, it is a waste as they are never really used. We fix it postponing the creation of the Visual Effect gameobject until it is required to render the preview.

To me there are still some optimizing opportunities if we were able to detect the Type of the editor associated to an asset, instead of having to create it to get the Type. That is outside the scope of VFX graph anyway