[2023.1] Improved HDR output shader keywords for less variants
When HDR Output display support landed for URP, it introduced many shader variants that needed to be stripped in URP and HDRP. There were 3 sets of keywords added to already keyword-heavy shaders (like FinalPost), making up to 18 combinations with 7 being invalid. To improve this, 2 set of keywords have been changed to dynamic branches using a constant buffer.

The 3 sets of keywords represented:
- what HDR operation to apply: None, Colorspace conversion, Color encoding, Both. (Shaders only defined some of them, except UberPost on URP.)
- what HDR colorspace to target: None, Rec709, Rec2020.
- what HDR encoding to target: None, Linear, PQ.

Now there's a single set of keyword for HDR operations (from 18 combinations to 4, all valid). The colorspaces and encodings are now ints and sent to the shaders as constants.