[HDRP] Fix issue with initilization of raytracing which could not be setup properly for few frames
Fix issue where when we setup raytracing application in HDRP, the code could have a null expection on the raytrace counter manager because it is not init but we still execute the code. The fix is to also affect the framesettings rayrtacing if the raytracing resources aren't properly initialized.

This fix an issue with the hdrp wizard where after clicking Fixing all, the DXR application was still not working.