Update URP PostPro test to GTF 7.12.0-exp.3
In #17000, we found that an update to XR Management (XRM) creating a settings entry with default settings caused the Graphics Testing Framework (GTF) to [incorrectly] believe that certain tests were being run in XR mode and to skip certain tests. Additionally, one of the filtered tests was the first test, and it had a side effect [intentionally used] for setting a fixed resolution for all of the remaining tests. Without the fixed resolution, a number of tests failed their image comparisons due to different aspect ratios.

This change updates to the fixed GTF package and an additional change to the URP tests to set resolution on _every_ test case, not just setting it on one and hoping that side effects work for the remaining tests.