[VFX] Fixed enum fields could be cropped if text is too long
Original PR: https://github.cds.internal.unity3d.com/unity/unity/pull/19993
https://jira.unity3d.com/browse/UUM-4420

Settings with dropdown are cropped in the UI making them unreadable.

Expected: Dropdown are not cropped and actual value is fully readable from the node UI

Tested on: 2022.2a16 but it's older than that. Not sure till when and if it's a regression.

![](https://jira.unity3d.com/secure/attachment/483030/483030_image-2022-06-08-17-36-51-659.png)