[HDRP][Water] Underwater node
Adds a SG node that can be used in post processes or post process custom pass to determine if a pixel is underwater and it's distance to the waterline.

The buffer cannot be fetched by script because it's allocated by rendergraph and thus we don't control it's lifetime, so we can't expose that to the user.

To get reasonnable precision on the distance value, we don't compute a fullscreen buffer anymore, we keep the 1D water line buffer and recompute the distance everytime we need it. This adds a few extra ALU at each sample but saves the 2D propagation pass (so essentially it just moves the ALU to somewhere else, with less bandwidth issue so it's actually faster), and i'd say most importantly we save a fullscreen texture allocation.

Last thing, reworked a bit the 1D propagation and it should remove some artifacts when looking at weird waterlines.

Example custom pass:

| Without | With |
| --- | --- |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/54d498bf-3a56-4a21-b8b9-bccd18cc3cc5) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/86b0f8f0-cf53-4ca7-8f08-ae8e9932265f) |

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/60bfeee6-58ce-4bc1-a169-930d03406782)