Sg/imgfixup
Some of our rotten tests were reenabled or test images out of date due to changes in SRPs. This PR should bring everything back into green that was out of date. It does not address _all_ failures, of which there are two known legitimate issues.

On Scarlett, our `IsNaNv4` test fails (LogicNodes)- this is a known failure that will need to be addressed separately.