[2023.1] [URP] Adding HDR color grading variant when HDR output is allowed
When HDR output is enabled, the only supported Color Grading Mode is High Dynamic Range. The render pipeline will override the color grading mode at runtime if HDR output is enabled, but the `_HDR_GRADING` variant of the UberPost shader may not have been included in the build. That resulted in the Uber PostProcessing saturating the image, and the result was too contrasted.

To fix this, the shader build pre-processor now checks if HDR output is allowed before stripping the `_HDR_GRADING` keyword.