[RPW]Fix reference sharing of AnimCurveParameters in the Volume Stack
The AnimCurve reference is being shared amongst the default values on the VolumeManager and the array that we are storing for fast access into the VolumeStack.

Avoid modifications of the Default State of the Volume Manager makes the Revert Property for AnimCurves to work as expected.

Also the values on the `VolumeStack `were being shared, making that the default state to be loss when interpolating any new value on it.

Issue found on this JIRA https://jira.unity3d.com/browse/UUM-20458 and this https://jira.unity3d.com/browse/UUM-20456