Uitech/rework/dispatch events refactor 2
Jira
- issue: https://jira.unity3d.com/browse/UIE-965
- epic: https://jira.unity3d.com/browse/UIE-871

This PR simplifies event dispatching behavior and API for UI Toolkit editor and runtime.
- removed (obsoleted) ExecuteDefaultAction and ExecuteDefaultActionAtTarget
- removed PropagationPhase AtTarget, DefaultAction, and DefaultActionAtTarget
- removed (obsoleted) PreventDefault method; all event prevention is now done with StopPropagation
- propagate events to composite roots exactly like regular non-root elements (no special AtTarget rule for events that wouldn't otherwise propagate)
- added HandleEventTrickleDown and HandleEventBubbleUp virtual methods to replace ExecuteDefaultAction
- added FocusController.IgnoreEvent method, to allow users to prevent focus changes caused by PointerDownEvent or NavigationMoveEvent; to avoid breaking too much existing code, focus change is happening during PostProcess of these event types, and calling StopPropagation at any point doesn't prevent it
- to stop ContextualMenuPopulateEvent with a prepopulated ContextualMenu from showing up, users can no longer call PreventDefault, and calling StopPropagation won't prevent the menu from showing up; however, users call call menu.Clear() and then StopImmediatePropagation to achieve the same result
- FocusEvent and BlurEvent are now sent multiple times per focus change, once per focusedElement being newly focused or blurred; this allows composite roots such as TextField to continue registering callbacks to those event types for the BubbleUp phase despite the leaf focused element being their TextElement child.
- MouseMoveEvent, MouseDownEvent, and MouseUpEvent compatibility events are now propagated in lock-step with the Pointer event that created them, and stopping either the Mouse or the Pointer event's propagation will now stop the propagation for both events. This should make it much easier to upgrade code based on mouse events to pointer events instead, which is highly desirable for code reusability between the Editor and Runtime, which has to support touch and pen input in addition to the mouse.

Code migration usually consists of replacing ExecuteDefaultAction and ExecuteDefaultActionAtTarget by HandleEventBubbleUp and PreventDefault for StopPropagation (or removing calls to PreventDefault if StopPropagation was already called in the same block of code). Using these simple rules work for 90% of the cases. In most of the rest of cases, the problem arises from the old code calling PreventDefault during a BubbleUp callback, which is no longer possible and can't be replaced by StopPropagation because the event has already reached its target. In that case, adding a callback on the TrickleDown phase to call StopPropagation is usually sufficient. There are the rare cases where neither of those two changes are enough to keep old code working. In those cases, a more careful case-by-case analysis is required, and the fix may not always be straightforward. For this reason, this PR should be considered heavily intrusive, and since the work was deemed desirable, it should land sooner than later to have some room to react to important regressions or criticism.

Problems will mostly arrive in situations where a lot of different callbacks from different systems interact and interleave at intricate moments during the propagation process. This PR basically unknots some of the crossing wires of UI Toolkit event propagation, and makes it so that, hopefully, there will be fewer ways to implement the same behavior in the new system, making it less complicated to grasp the big picture when reading new code, and leaving less room for complicated hackly solutions.

An example of event propagation
A
|
B (composite root)
|
C (target)


The old propagation rules
- TrickleDown phase:
  - if !isPropagationStopped, invoke A's TrickleDown callbacks
- AtTarget phase:
  - if !isPropagationStopped, invoke C's TrickleDown callbacks
  - if !isPropagationStopped, invoke C's BubbleUp callbacks
  - if !isPropagationStopped, invoke B's TrickleDown callbacks
  - if !isPropagationStopped, invoke B's BubbleUp callbacks
- DefaultActionAtTarget phase:
  - if !isDefaultPrevented, call C's ExecuteDefaultActionAtTarget method override
  - if !isDefaultPrevented, call B's ExecuteDefaultActionAtTarget method override
- BubbleUp phase
  - if !isPropagationStopped, invoke A's BubbleUp callbacks
- DefaultAction phase
  - if !isDefaultPrevented, call C's ExecuteDefaultAction method override
  - if !isDefaultPrevented, call B's ExecuteDefaultAction method override

The new propagation rules
- TrickleDown phase:
  - if !isPropagationStopped, invoke A's TrickleDown callbacks then call A's HandleEventTrickleDown method override
  - if !isPropagationStopped, invoke B's TrickleDown callbacks then call B's HandleEventTrickleDown method override
  - if !isPropagationStopped, invoke C's TrickleDown callbacks then call C's HandleEventTrickleDown method override
- BubbleUp phase:
  - if !isPropagationStopped, call C's HandleEventBubbleUp method override then invoke C's BubbleUp callbacks
  - if !isPropagationStopped, call B's HandleEventBubbleUp method override then invoke B's BubbleUp callbacks
  - if !isPropagationStopped, call A's HandleEventBubbleUp method override then invoke A's BubbleUp callbacks