Fix for LOD crossfade in URP when using BatchRendererGroup
Shader fixes to the DOTS instancing variants so that URP LOD crossfade works correctly when using BatchRendererGroup.

In this variant, the cross-fade value is read from a per-instance parameter, so the shader variant needs to propagate the instance ID to the pixel shader.  Most of the changes are to fix some missing macros to pass the instance ID.