[VFX] Hide Exposure Weight slider when a shadergraph is assigned
Jira:https://jira.unity3d.com/browse/UUM-23614

Repro steps:
1. In HDRP, create a new Unlit ShaderGraph
2. Open it and enable VFX Graph (new integration)
3. Create a new VFX, select its output and in the inspector enable 'Use Exposure Weight'
4. Assign the SG to the output.
5. Observe the exposure weight slider is shown

Expected outcome:

The exposure weight slider to disappear if a ShaderGraph is used
Notes:

The 'Use Exposure Weight' button is correctly hidden if the SG is assigned first. This is just the extra step of removing the slider if it was displayed before the SG was assigned.